feat:1.4L CP6 云盘-当用户未加入一个团队时正确显示错误结果

+++ src/main/java/com/ksptooi/biz/drive/controller/EntryController.java
@@ -1,5 +1,6 @@
 package com.ksptooi.biz.drive.controller;
 
+import com.ksptooi.biz.core.service.AuthService;
 import com.ksptooi.biz.drive.model.dto.*;
 import com.ksptooi.biz.drive.model.vo.CheckEntryMoveVo;
 import com.ksptooi.biz.drive.model.vo.GetDriveInfo;
@@ -8,7 +9,6 @@ import com.ksptooi.biz.drive.model.vo.GetEntryListVo;
 import com.ksptooi.biz.drive.service.EntryService;
 import com.ksptooi.commons.annotation.PrintLog;
 import com.ksptool.assembly.entity.web.CommonIdDto;
-import com.ksptool.assembly.entity.web.PageResult;
 import com.ksptool.assembly.entity.web.Result;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -33,12 +33,22 @@ public class EntryController {
     @PostMapping("/getDriveInfo")
     @Operation(summary = "获取云盘信息")
     public Result<GetDriveInfo> getDriveInfo() throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         return Result.success(entryService.getDriveInfo());
     }
 
     @PostMapping("/getEntryList")
     @Operation(summary = "查询条目列表")
     public Result<GetEntryListVo> getEntryList(@RequestBody @Valid GetEntryListDto dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         var ret = entryService.getEntryList(dto);
         return Result.success(ret);
     }
@@ -46,6 +56,10 @@ public class EntryController {
     @Operation(summary = "新增条目")
     @PostMapping("/addEntry")
     public Result<String> addEntry(@RequestBody @Valid AddEntryDto dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
 
         //验证输入参数
         if (dto.validate() != null) {
@@ -60,6 +74,11 @@ public class EntryController {
     @Operation(summary = "复制条目")
     @PostMapping("/copyEntry")
     public Result<String> copyEntry(@RequestBody @Valid CopyEntryDto dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         entryService.copyEntry(dto);
         return Result.success("复制成功");
     }
@@ -67,6 +86,11 @@ public class EntryController {
     @Operation(summary = "重命名条目")
     @PostMapping("/renameEntry")
     public Result<String> renameEntry(@RequestBody @Valid RenameEntry dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         entryService.renameEntry(dto);
         return Result.success("重命名成功");
     }
@@ -74,6 +98,11 @@ public class EntryController {
     @Operation(summary = "移动检测")
     @PostMapping("/checkEntryMove")
     public Result<CheckEntryMoveVo> checkEntryMove(@RequestBody @Valid MoveEntryDto dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         var ret = entryService.checkEntryMove(dto);
         return Result.success(ret);
     }
@@ -82,6 +111,11 @@ public class EntryController {
     @Operation(summary = "移动条目")
     @PostMapping("/moveEntry")
     public Result<String> moveEntry(@RequestBody @Valid MoveEntryDto dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         entryService.moveEntry(dto);
         return Result.success("移动成功");
     }
@@ -89,6 +123,11 @@ public class EntryController {
     @Operation(summary = "查询条目详情")
     @PostMapping("/getEntryDetails")
     public Result<GetEntryDetailsVo> getEntryDetails(@RequestBody @Valid CommonIdDto dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         GetEntryDetailsVo details = entryService.getEntryDetails(dto);
         if (details == null) {
             return Result.error("无数据");
@@ -99,6 +138,11 @@ public class EntryController {
     @Operation(summary = "删除条目")
     @PostMapping("/removeEntry")
     public Result<String> removeEntry(@RequestBody @Valid CommonIdDto dto) throws Exception {
+
+        if (AuthService.getCurrentCompanyId() == null) {
+            return Result.error(101, "该操作需要用户加入团队后才能执行");
+        }
+
         entryService.removeEntry(dto);
         return Result.success("操作成功");
     }
